Add is_partial_ok() function to restrict the push down condition.
Check that partial aggregate agg is fine to push down.
condition1) agg is AGGKIND_NORMAL aggregate which contains no distinct
  or order by clauses
condition2) The aggtranstype is not internal and there is no aggfinalfn:
  in this case, nothing we need to do.
condition3) The aggtranstype is not internal, but aggfinalfn is not null:
  in this case, we need to deparse the interal result into the remote sql.
condition4) The aggtranstype is internal:
  We may need to do a deparse, and then we need to do a trans-code to
  transform the final-result from remote server into partial-result.

And we will support more agg functions in the future, so add one
supported array to control the condition.