Adjust 'numsegments' of policy for foreign table.
Make 'numsegments' of distribution policy for foreign table same
as the foreign server option 'num_segments'.

If the numsegments of gppolicy for foreign table is same as the GPDB
segment number, then orca optimizer will generate a wrong plan for scanning
for a foreign table when the default segment number is not equal to the
foreign server option 'num_segments'.

And orca optimizer does not support partially distributed tables yet, so
in this case, the planner will fallback to postgres optimizer when foreign
server option 'num_segments' is not equal to the default GPDB segment number.